Add files via upload
Debugged checker methods in Board UseCase, added feature for King Entities to move in a castling fashion. Debugged and fixed MoveBuffer UseCase and GameActivity Controller to prevent the program from crashing whenever the King Entities attempted to move towards a square that put them in checkmate. Included package-protected getCopy method for all chess pieces, to allow the Board class to generate a shallow copy of a board at a given state.